+++ packages/components/src/form/RadioGroup/RadioGroup.js
@@ -1,6 +1,6 @@
 import React, { forwardRef, useEffect, useMemo, useRef, useState } from 'react';
 import PropTypes from 'prop-types';
-import { isEmpty, map } from 'lodash';
+import { isEmpty, map, noop } from 'lodash';
 import { Box, SegmentedControl as MantineSegmentedControl } from '@mantine/core';
 import { useId } from '@mantine/hooks';
 import { RadioGroupStyles } from './RadioGroup.styles';
@@ -11,7 +11,6 @@ import {
   INPUT_WRAPPER_SIZES,
   InputWrapper,
 } from '../InputWrapper';
-import { isFunction } from 'lodash';
 
 export const RADIOGROUP_DIRECTIONS = ['column', 'row'];
 
@@ -66,9 +65,10 @@ const RadioGroup = forwardRef(
       imageHeight,
       fullWidth,
       useAria,
+      onChange = noop,
       ...props
     },
-    ref
+    ref,
   ) => {
     const [value, setValue] = useState(props.value);
     const [activePosition, setActivePosition] = useState({ height: 0, translate: 0 });
@@ -83,7 +83,7 @@ const RadioGroup = forwardRef(
 
     const { classes, cx } = RadioGroupStyles(
       { variant, value, direction, fullWidth, activePosition, hasError, rounded },
-      { name: 'RadioGroup' }
+      { name: 'RadioGroup' },
     );
 
     if (!direction) {
@@ -97,11 +97,11 @@ const RadioGroup = forwardRef(
       direction = 'row';
     }
 
-    const onChange = (value) => {
-      const item = data.find((item) => item.value === value);
-      if (!props.disabled && !item.disabled) {
-        isFunction(props.onChange) && props.onChange(value);
-        setValue(value);
+    const handleOnChange = (val) => {
+      const option = data.find((item) => item.value === val);
+      if (!props.disabled && !option.disabled) {
+        onChange(val);
+        setValue(val);
       }
     };
 
@@ -141,11 +141,11 @@ const RadioGroup = forwardRef(
             ref={ref}
             id={uuid}
             size={size}
-            onChange={onChange}
+            onChange={handleOnChange}
             classNames={classes}
-            defaultValue={defaultValue ? defaultValue : ' '}
+            defaultValue={defaultValue || ' '}
             value={value}
-            data={map(data, ({ label, ...item }, index) => ({
+            data={map(data, ({ label: itemLabel, ...item }, index) => ({
               value: item.value,
               label: (
                 <Radio
@@ -163,7 +163,7 @@ const RadioGroup = forwardRef(
                   imageHeight={imageHeight}
                   useAria={useAria}
                 >
-                  {label}
+                  {itemLabel}
                 </Radio>
               ),
             }))}
@@ -171,9 +171,10 @@ const RadioGroup = forwardRef(
         </Box>
       </InputWrapper>
     );
-  }
+  },
 );
 
+RadioGroup.displayName = 'RadioGroup';
 RadioGroup.propTypes = RADIOGROUP_PROP_TYPES;
 RadioGroup.defaultProps = RADIOGROUP_DEFAULT_PROPS;