+++ packages/leemons/src/common/AssetPlayer/components/AudioCardPlayer/AudioCardPlayer.constants.js
@@ -1,4 +1,0 @@
-import PropTypes from 'prop-types';
-
-export const AUDIO_CARD_PLAYER_DEFAULT_PROPS = {};
-export const AUDIO_CARD_PLAYER_PROP_TYPES = {};