+++ packages/leemons/src/common/AssetPlayer/components/AudioProgressBar/index.js
@@ -1,2 +1,0 @@
-export * from './AudioProgressBar';
-export * from './AudioProgressBar.constants';