+++ packages/leemons/src/common/AssetPlayer/index.js
@@ -1,2 +1,0 @@
-export * from './AssetPlayer';
-export * from './AssetPlayer.constants';